Adding new parameter package_ensure
This allows users to configure the desired state of Package['metricbeat']
when $ensure == 'present'. Allowing them to ensure that the latest available
package is installed, or a specific version.

Possible issue: There is no easy way to determine an absent type value
with this parameter. The current recommendation for removing the package
from a node is to set $ensure to absent.