Adding configuration file validation and the ability to disable it.
New parameter 'disable_configtest' grants the option to disable
File['metricbeat.yml'] validation. Defaults to false and it is
recommended to remain false.

When File['metricbeat.yml'] changes a validation command is executed
before replacing the new file. This simply ensures that syntax is
correct and is a layer of protection from unexpected failures.